refactor: 예약 생성 API 요청 DTO 분리
