+++ src/main/java/roomescape/domain/Reservation.java
@@ -4,41 +4,21 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class Reservation {
-    
-    public static final String ERROR_MESSAGE_FORMAT = "[ERROR]  예약 필수 정보가 누락되었습니다. %s: %s";
 
     private Long id;
     private String name;
     private LocalDate date;
     private LocalTime time;
 
+    public Reservation(String name, LocalDate date, LocalTime time) {
+        this(null, name, date, time);
+    }
 
     public Reservation(Long id, String name, LocalDate date, LocalTime time) {
-        validateName(name);
-        validateDate(date);
-        validateTime(time);
         this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
-    }
-
-    private void validateName(String name) {
-        if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException(String.format(ERROR_MESSAGE_FORMAT, "name", name));
-        }
-    }
-
-    private void validateDate(LocalDate date) {
-        if (date == null) {
-            throw new IllegalArgumentException(String.format(ERROR_MESSAGE_FORMAT, "date", date));
-        }
-    }
-
-    private void validateTime(LocalTime time) {
-        if (time == null) {
-            throw new IllegalArgumentException(String.format(ERROR_MESSAGE_FORMAT, "time", time));
-        }
     }
 
     public Long getId() {
